Inverse Path Distance Weighting

Interpolate geo-referenced point data using inverse path distance weighting.

ipdw(
  sf_ob,
  costras,
  range,
  paramlist,
  overlapped = FALSE,
  yearmon = "default",
  removefile = TRUE,
  step = 16,
  dist_power = 1,
  trim_rstack = FALSE
)

Arguments

sf_ob

sf object with point geometries

costras

RasterLayer. Cost raster

range

numeric. Range of interpolation neighborhood

paramlist

character. String representing parameter names

overlapped

logical. Default is FALSE, specify TRUE if some points lie on top of barriers

yearmon

character. String specifying the name of the sf_ob

removefile

logical. Remove files after processing?

step

numeric. Number of sub loops to manage memory during raster processing.

dist_power

numeric. Distance decay power (p)

trim_rstack

logical. Trim the raster output by the convex hill of sf_ob

Value

RasterLayer

Details

This is a high level function that interpolates an sf object with point geometries in a single pass.

Points must be located within a single contiguous area. The presence of "landlocked" points will cause errors. It may be necessary to increase the value assigned to land areas when using a large range value in combination with a large sized cost rasters (grain x extent). In these cases, the value of land areas should be increased to ensure that it is always greater than the maximum accumulated cost path distance of any given geo-referenced point.

Examples

# see vignette